[SA - statistic overview] update widget query dynamic parameter

+++ features/merchant/seller_home_common/src/main/java/com/tokopedia/sellerhomecommon/presentation/model/LineGraphDataUiModel.kt
@@ -16,5 +16,5 @@ data class LineGraphDataUiModel(
 data class XYAxisUiModel(
         val xLabel: String,
         val yLabel: String,
-        val yVal: Int
+        val yVal: Float
 )
